+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_statistical.py
@@ -1,0 +1,488 @@
+# global
+from hypothesis import strategies as st
+import numpy as np
+
+# local
+import ivy_tests.test_ivy.helpers as helpers
+import ivy_tests.test_ivy.test_frontends.test_numpy.helpers as np_helpers
+from ivy_tests.test_ivy.helpers import handle_frontend_test
+from ivy_tests.test_ivy.test_functional.test_core.test_statistical import (
+    statistical_dtype_values,
+    _get_castable_dtype,
+)
+
+
+# einsum
+@handle_frontend_test(
+    fn_tree="jax.numpy.einsum",
+    eq_n_op=st.sampled_from(
+        [
+            (
+                "ii",
+                np.arange(25).reshape(5, 5),
+            ),
+            (
+                "ii->i",
+                np.arange(25).reshape(5, 5),
+            ),
+            ("ij,j", np.arange(25).reshape(5, 5), np.arange(5)),
+        ]
+    ),
+    dtype=helpers.get_dtypes("float", full=False),
+)
+def test_jax_numpy_einsum(
+    *,
+    eq_n_op,
+    dtype,
+    as_variable,
+    native_array,
+    with_out,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    kw = {}
+    i = 0
+    for x_ in eq_n_op:
+        kw["x{}".format(i)] = x_
+        i += 1
+    num_positional_args = i
+    helpers.test_frontend_function(
+        input_dtypes=dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        **kw,
+        out=None,
+        optimize="optimal",
+        precision=None,
+        _use_xeinsum=False,
+    )
+
+
+# mean
+@handle_frontend_test(
+    fn_tree="jax.numpy.mean",
+    dtype_x_axis=statistical_dtype_values(function="mean"),
+    dtype=helpers.get_dtypes("float", full=False, none=True),
+    where=np_helpers.where(),
+    keepdims=st.booleans(),
+)
+def test_jax_numpy_mean(
+    *,
+    dtype_x_axis,
+    dtype,
+    keepdims,
+    where,
+    num_positional_args,
+    with_out,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    x_dtype, x, axis = dtype_x_axis
+    if isinstance(axis, tuple):
+        axis = axis[0]
+    where, as_variable, native_array = np_helpers.handle_where_and_array_bools(
+        where=where,
+        input_dtype=x_dtype,
+        as_variable=as_variable,
+        native_array=native_array,
+    )
+
+    np_helpers.test_frontend_function(
+        input_dtypes=x_dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        atol=1e-2,
+        rtol=1e-2,
+        a=x[0],
+        axis=axis,
+        dtype=dtype[0],
+        out=None,
+        keepdims=keepdims,
+        where=where,
+    )
+
+
+# var
+@handle_frontend_test(
+    fn_tree="jax.numpy.var",
+    dtype_x_axis=statistical_dtype_values(function="var"),
+    dtype=helpers.get_dtypes("float", full=False, none=True),
+    where=np_helpers.where(),
+    keepdims=st.booleans(),
+)
+def test_jax_numpy_var(
+    *,
+    dtype_x_axis,
+    dtype,
+    keepdims,
+    where,
+    num_positional_args,
+    with_out,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    x_dtype, x, axis, ddof = dtype_x_axis
+    if isinstance(axis, tuple):
+        axis = axis[0]
+    where, as_variable, native_array = np_helpers.handle_where_and_array_bools(
+        where=where,
+        input_dtype=x_dtype,
+        as_variable=as_variable,
+        native_array=native_array,
+    )
+
+    np_helpers.test_frontend_function(
+        input_dtypes=x_dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        a=x[0],
+        axis=axis,
+        dtype=dtype[0],
+        out=None,
+        ddof=ddof,
+        keepdims=keepdims,
+        where=where,
+    )
+
+
+# argmin
+@handle_frontend_test(
+    fn_tree="jax.numpy.argmin",
+    dtype_and_x=helpers.dtype_values_axis(
+        available_dtypes=helpers.get_dtypes("numeric"),
+        force_int_axis=True,
+        min_num_dims=1,
+        valid_axis=True,
+    ),
+    keepdims=st.booleans(),
+)
+def test_jax_numpy_argmin(
+    *,
+    dtype_and_x,
+    keepdims,
+    as_variable,
+    with_out,
+    num_positional_args,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    input_dtype, x, axis = dtype_and_x
+    helpers.test_frontend_function(
+        input_dtypes=input_dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        a=x[0],
+        axis=axis,
+        out=None,
+        keepdims=keepdims,
+    )
+
+
+# bincount
+@handle_frontend_test(
+    fn_tree="jax.numpy.bincount",
+    dtype_and_x=helpers.dtype_and_values(
+        available_dtypes=helpers.get_dtypes("integer"),
+        min_value=1,
+        max_value=2,
+        shape=st.shared(
+            helpers.get_shape(
+                min_num_dims=1,
+                max_num_dims=1,
+            ),
+            key="a_s_d",
+        ),
+    ),
+)
+def test_jax_numpy_bincount(
+    *,
+    dtype_and_x,
+    num_positional_args,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    input_dtype, x = dtype_and_x
+    helpers.test_frontend_function(
+        input_dtypes=input_dtype,
+        as_variable_flags=as_variable,
+        with_out=False,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        x=x[0],
+        weights=None,
+        minlength=0,
+        length=None,
+    )
+
+
+# cumprod
+@handle_frontend_test(
+    fn_tree="jax.numpy.cumprod",
+    dtype_x_axis=helpers.dtype_values_axis(
+        available_dtypes=helpers.get_dtypes("numeric"),
+        min_num_dims=1,
+        max_num_dims=5,
+        min_value=-100,
+        max_value=100,
+        valid_axis=True,
+        allow_neg_axes=False,
+        max_axes_size=1,
+        force_int_axis=True,
+    ),
+    dtype=helpers.get_dtypes("float", none=True, full=False),
+)
+def test_jax_numpy_cumprod(
+    *,
+    dtype_x_axis,
+    dtype,
+    num_positional_args,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    input_dtype, x, axis = dtype_x_axis
+    helpers.test_frontend_function(
+        input_dtypes=input_dtype,
+        as_variable_flags=as_variable,
+        with_out=False,
+        all_aliases=["numpy.cumproduct"],
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        rtol=1e-2,
+        a=x[0],
+        axis=axis,
+        dtype=dtype[0],
+    )
+
+
+# cumsum
+@handle_frontend_test(
+    fn_tree="jax.numpy.cumsum",
+    dtype_x_axis=helpers.dtype_values_axis(
+        available_dtypes=helpers.get_dtypes("numeric"),
+        min_num_dims=1,
+        max_num_dims=5,
+        min_value=-100,
+        max_value=100,
+        valid_axis=True,
+        allow_neg_axes=False,
+        max_axes_size=1,
+        force_int_axis=True,
+    ),
+    dtype=helpers.get_dtypes("numeric", none=True, full=False),
+)
+def test_jax_numpy_cumsum(
+    *,
+    dtype_x_axis,
+    dtype,
+    num_positional_args,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    input_dtype, x, axis = dtype_x_axis
+    helpers.test_frontend_function(
+        input_dtypes=input_dtype,
+        as_variable_flags=as_variable,
+        with_out=False,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        a=x[0],
+        axis=axis,
+        dtype=dtype[0],
+    )
+
+
+# sum
+@handle_frontend_test(
+    fn_tree="jax.numpy.sum",
+    dtype_x_axis_castable=_get_castable_dtype(),
+    initial=st.none() | st.floats(-10.0, 10.0),
+    where=np_helpers.where(),
+    keepdims=st.booleans(),
+)
+def test_jax_numpy_sum(
+    *,
+    dtype_x_axis_castable,
+    initial,
+    where,
+    keepdims,
+    with_out,
+    num_positional_args,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    x_dtype, x, axis, castable_dtype = dtype_x_axis_castable
+
+    if isinstance(axis, tuple):
+        axis = axis[0]
+    where, as_variable, native_array = np_helpers.handle_where_and_array_bools(
+        where=where,
+        input_dtype=x_dtype,
+        as_variable=as_variable,
+        native_array=native_array,
+    )
+
+    np_helpers.test_frontend_function(
+        input_dtypes=[x_dtype],
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        rtol=1e-1,
+        atol=1e-2,
+        a=x[0],
+        axis=axis,
+        dtype=castable_dtype,
+        out=None,
+        keepdims=keepdims,
+        initial=initial,
+        where=where,
+    )
+
+
+# min
+@handle_frontend_test(
+    fn_tree="jax.numpy.min",
+    dtype_x_axis=statistical_dtype_values(function="min"),
+    where=np_helpers.where(),
+    keepdims=st.booleans(),
+)
+def test_jax_numpy_min(
+    *,
+    dtype_x_axis,
+    keepdims,
+    where,
+    num_positional_args,
+    with_out,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    x_dtype, x, axis = dtype_x_axis
+    if isinstance(axis, tuple):
+        axis = axis[0]
+    where, as_variable, native_array = np_helpers.handle_where_and_array_bools(
+        where=where,
+        input_dtype=x_dtype,
+        as_variable=as_variable,
+        native_array=native_array,
+    )
+
+    np_helpers.test_frontend_function(
+        input_dtypes=x_dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        all_aliases=["numpy.amin"],
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        a=x[0],
+        axis=axis,
+        out=None,
+        keepdims=keepdims,
+        where=where,
+    )
+
+
+# max
+@handle_frontend_test(
+    fn_tree="jax.numpy.max",
+    dtype_x_axis=statistical_dtype_values(function="max"),
+    where=np_helpers.where(),
+    keepdims=st.booleans(),
+)
+def test_jax_numpy_max(
+    *,
+    dtype_x_axis,
+    keepdims,
+    where,
+    num_positional_args,
+    with_out,
+    as_variable,
+    native_array,
+    on_device,
+    fn_tree,
+    frontend,
+):
+    x_dtype, x, axis = dtype_x_axis
+    if isinstance(axis, tuple):
+        axis = axis[0]
+    where, as_variable, native_array = np_helpers.handle_where_and_array_bools(
+        where=where,
+        input_dtype=x_dtype,
+        as_variable=as_variable,
+        native_array=native_array,
+    )
+
+    np_helpers.test_frontend_function(
+        input_dtypes=x_dtype,
+        as_variable_flags=as_variable,
+        with_out=with_out,
+        all_aliases=["numpy.amax"],
+        num_positional_args=num_positional_args,
+        native_array_flags=native_array,
+        frontend=frontend,
+        fn_tree=fn_tree,
+        on_device=on_device,
+        a=x[0],
+        axis=axis,
+        out=None,
+        keepdims=keepdims,
+        where=where,
+    )
